+++ ATM.DAL/Models/Admin.cs
@@ -5,7 +5,7 @@ namespace ATM.DAL.Models
     internal class Admin : User
     {
 
-        public Admin(long id, string password) : base(id, password)
+        public Admin(int id, string password) : base(id, password)
         {
         }
