+++ ATM.DAL/Models/Customer.cs
@@ -4,7 +4,7 @@ namespace ATM.DAL.Models
 {
     public class Customer : User
     {
-        public Customer(long id, string password) : base(id, password) { }
+        public Customer(int id, string password) : base(id, password) { }
         public override Role Role { get; } = Role.Customer;
         public Account Account { get; set; }
     }
