+++ ATM.DAL/Models/ThirdParty.cs
@@ -4,7 +4,7 @@ namespace ATM.DAL.Models
 {
     internal class ThirdParty : User
     {
-        public ThirdParty(long id, string password) : base(id, password) { }
+        public ThirdParty(int id, string password) : base(id, password) { }
         public override Role Role { get; } = Role.ThirdParty;
         public Account Account { get; set; }
     }
